Challenge View upgrades
Chevron buttons added to the search field that help the user navigate through the list of results for a given search query. Counter written and added to view. Challenge destination title also displayed in view. Overlays that hold the counter, title and click counter all had their background properties adjusted so that they fit only around the content and dont extend to the bottom of the screen.